@@ -17,72 +17,36 @@
     </modules>
 
 	<properties>
-		<javax.servlet-api.version>3.1.0</javax.servlet-api.version>
-		<javax.servlet.jsp-api.version>2.3.3</javax.servlet.jsp-api.version>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+		<maven.compiler.encoding>UTF-8</maven.compiler.encoding>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.test.skip>true</maven.test.skip>
 
-		<spring.version>4.3.19.RELEASE</spring.version>
-		<jackson.version>2.9.6</jackson.version>
-		<aspectjweaver.version>1.8.13</aspectjweaver.version>
-		<slf4j-api.version>1.7.25</slf4j-api.version>
-		<freemarker.version>2.3.28</freemarker.version>
-		<junit.version>4.12</junit.version>
+		<netty-all.version>4.1.58.Final</netty-all.version>
+		<gson.version>2.8.6</gson.version>
 
-		<jetty-server.version>9.2.26.v20180806</jetty-server.version>
-		<hessian.version>4.0.51</hessian.version>
-		<httpclient.version>4.5.6</httpclient.version>
-		
-		<commons-exec.version>1.3</commons-exec.version>
-		<commons-collections4.version>4.2</commons-collections4.version>
-		<commons-lang3.version>3.8.1</commons-lang3.version>
-		<commons-email.version>1.5</commons-email.version>
+		<spring.version>5.3.3</spring.version>
+		<spring-boot.version>2.4.2</spring-boot.version>
 
-		<c3p0.version>0.9.5.2</c3p0.version>
-		<mysql-connector-java.version>5.1.47</mysql-connector-java.version>
-		<mybatis-spring.version>1.3.2</mybatis-spring.version>
-		<mybatis.version>3.4.6</mybatis.version>
+		<mybatis-spring-boot-starter.version>2.1.4</mybatis-spring-boot-starter.version>
+		<mysql-connector-java.version>8.0.23</mysql-connector-java.version>
 
-		<groovy-all.version>2.4.15</groovy-all.version>
-		<quartz.version>2.3.0</quartz.version>
+		<slf4j-api.version>1.7.30</slf4j-api.version>
+		<junit.version>5.7.1</junit.version>
+		<javax.annotation-api.version>1.3.2</javax.annotation-api.version>
 
-		<spring-boot.version>1.5.16.RELEASE</spring-boot.version>
+		<groovy.version>3.0.7</groovy.version>
+
+		<maven-source-plugin.version>3.2.1</maven-source-plugin.version>
+		<maven-javadoc-plugin.version>3.2.0</maven-javadoc-plugin.version>
+		<maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
+		<maven-war-plugin.version>3.3.1</maven-war-plugin.version>
 	</properties>
 
 	<build>
 		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.1</version>
-				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
-					<encoding>UTF8</encoding>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-resources-plugin</artifactId>
-				<version>2.6</version>
-				<configuration>
-					<encoding>UTF-8</encoding>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-war-plugin</artifactId>
-				<version>2.2</version>
-				<configuration>
-					<archiveClasses>false</archiveClasses>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-				<version>2.19.1</version>
-				<configuration>
-					<skip>true</skip>
-				</configuration>
-			</plugin>
 		</plugins>
 	</build>
 
