@@ -1,9 +1,8 @@
 package com.xxl.job.admin.controller;
 
-import com.xxl.job.admin.controller.interceptor.PermissionInterceptor;
-import com.xxl.job.admin.core.conf.XxlJobAdminConfig;
-import org.junit.Before;
-import org.junit.Test;
+import com.xxl.job.admin.service.LoginService;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.util.LinkedMultiValueMap;
@@ -17,21 +16,22 @@ public class JobInfoControllerTest extends AbstractSpringMvcTest {
 
   private Cookie cookie;
 
-  @Before
+  @BeforeEach
   public void login() throws Exception {
     MvcResult ret = mockMvc.perform(
         post("/login")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            .param("userName", XxlJobAdminConfig.getAdminConfig().getLoginUsername())
-            .param("password", XxlJobAdminConfig.getAdminConfig().getLoginPassword())
+            .param("userName", "admin")
+            .param("password", "123456")
     ).andReturn();
-    cookie = ret.getResponse().getCookie(PermissionInterceptor.LOGIN_IDENTITY_KEY);
+    cookie = ret.getResponse().getCookie(LoginService.LOGIN_IDENTITY_KEY);
   }
 
   @Test
   public void testAdd() throws Exception {
     MultiValueMap<String, String> parameters = new LinkedMultiValueMap<String, String>();
     parameters.add("jobGroup", "1");
+    parameters.add("triggerStatus", "-1");
 
     MvcResult ret = mockMvc.perform(
         post("/jobinfo/pageList")
