@@ -1,69 +1,72 @@
 package com.xxl.job.admin.controller;
 
-import com.xxl.job.admin.controller.annotation.PermessionLimit;
+import com.xxl.job.admin.controller.annotation.PermissionLimit;
+import com.xxl.job.admin.core.conf.XxlJobAdminConfig;
 import com.xxl.job.core.biz.AdminBiz;
-import com.xxl.job.core.rpc.codec.RpcRequest;
-import com.xxl.job.core.rpc.codec.RpcResponse;
-import com.xxl.job.core.rpc.netcom.NetComServerFactory;
-import com.xxl.job.core.rpc.serialize.HessianSerializer;
-import com.xxl.job.core.util.HttpClientUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.xxl.job.core.biz.model.HandleCallbackParam;
+import com.xxl.job.core.biz.model.RegistryParam;
+import com.xxl.job.core.biz.model.ReturnT;
+import com.xxl.job.core.util.GsonTool;
+import com.xxl.job.core.util.XxlJobRemotingUtil;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.OutputStream;
+import java.util.List;
 
 /**
  * Created by xuxueli on 17/5/10.
  */
 @Controller
+@RequestMapping("/api")
 public class JobApiController {
-    private static Logger logger = LoggerFactory.getLogger(JobApiController.class);
 
-    private RpcResponse doInvoke(HttpServletRequest request) {
-        try {
-            // deserialize request
-            byte[] requestBytes = HttpClientUtil.readBytes(request);
-            if (requestBytes == null || requestBytes.length==0) {
-                RpcResponse rpcResponse = new RpcResponse();
-                rpcResponse.setError("RpcRequest byte[] is null");
-                return rpcResponse;
-            }
-            RpcRequest rpcRequest = (RpcRequest) HessianSerializer.deserialize(requestBytes, RpcRequest.class);
+    @Resource
+    private AdminBiz adminBiz;
 
-            // invoke
-            RpcResponse rpcResponse = NetComServerFactory.invokeService(rpcRequest, null);
-            return rpcResponse;
-        } catch (Exception e) {
-            logger.error(e.getMessage(), e);
+    /**
+     * api
+     *
+     * @param uri
+     * @param data
+     * @return
+     */
+    @RequestMapping("/{uri}")
+    @ResponseBody
+    @PermissionLimit(limit=false)
+    public ReturnT<String> api(HttpServletRequest request, @PathVariable("uri") String uri, @RequestBody(required = false) String data) {
 
-            RpcResponse rpcResponse = new RpcResponse();
-            rpcResponse.setError("Server-error:" + e.getMessage());
-            return rpcResponse;
+        // valid
+        if (!"POST".equalsIgnoreCase(request.getMethod())) {
+            return new ReturnT<String>(ReturnT.FAIL_CODE, "invalid request, HttpMethod not support.");
         }
-    }
+        if (uri==null || uri.trim().length()==0) {
+            return new ReturnT<String>(ReturnT.FAIL_CODE, "invalid request, uri-mapping empty.");
+        }
+        if (XxlJobAdminConfig.getAdminConfig().getAccessToken()!=null
+                && XxlJobAdminConfig.getAdminConfig().getAccessToken().trim().length()>0
+                && !XxlJobAdminConfig.getAdminConfig().getAccessToken().equals(request.getHeader(XxlJobRemotingUtil.XXL_JOB_ACCESS_TOKEN))) {
+            return new ReturnT<String>(ReturnT.FAIL_CODE, "The access token is wrong.");
+        }
 
-    @RequestMapping(AdminBiz.MAPPING)
-    @PermessionLimit(limit=false)
-    public void api(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        // services mapping
+        if ("callback".equals(uri)) {
+            List<HandleCallbackParam> callbackParamList = GsonTool.fromJson(data, List.class, HandleCallbackParam.class);
+            return adminBiz.callback(callbackParamList);
+        } else if ("registry".equals(uri)) {
+            RegistryParam registryParam = GsonTool.fromJson(data, RegistryParam.class);
+            return adminBiz.registry(registryParam);
+        } else if ("registryRemove".equals(uri)) {
+            RegistryParam registryParam = GsonTool.fromJson(data, RegistryParam.class);
+            return adminBiz.registryRemove(registryParam);
+        } else {
+            return new ReturnT<String>(ReturnT.FAIL_CODE, "invalid request, uri-mapping("+ uri +") not found.");
+        }
 
-        // invoke
-        RpcResponse rpcResponse = doInvoke(request);
-
-        // serialize response
-        byte[] responseBytes = HessianSerializer.serialize(rpcResponse);
-
-        response.setContentType("text/html;charset=utf-8");
-        response.setStatus(HttpServletResponse.SC_OK);
-        //baseRequest.setHandled(true);
-
-        OutputStream out = response.getOutputStream();
-        out.write(responseBytes);
-        out.flush();
     }
 
 }
