@@ -12,7 +12,7 @@ import javax.servlet.http.HttpServletResponse;
 public class CookieUtil {
 
 	// 默认缓存时间,单位/秒, 2H
-	private static final int COOKIE_MAX_AGE = 60 * 60 * 2;
+	private static final int COOKIE_MAX_AGE = Integer.MAX_VALUE;
 	// 保存路径,根路径
 	private static final String COOKIE_PATH = "/";
 	
