@@ -1,8 +1,6 @@
 package com.xxl.job.admin.core.util;
 
 import com.xxl.job.admin.core.conf.XxlJobAdminConfig;
-import com.xxl.job.core.util.JacksonUtil;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ClassPathResource;
@@ -32,8 +30,7 @@ public class I18nUtil {
         try {
             // build i18n prop
             String i18n = XxlJobAdminConfig.getAdminConfig().getI18n();
-            i18n = StringUtils.isNotBlank(i18n)?("_"+i18n):i18n;
-            String i18nFile = MessageFormat.format("i18n/message{0}.properties", i18n);
+            String i18nFile = MessageFormat.format("i18n/message_{0}.properties", i18n);
 
             // load prop
             Resource resource = new ClassPathResource(i18nFile);
@@ -62,7 +59,7 @@ public class I18nUtil {
      * @return
      */
     public static String getMultString(String... keys) {
-        Map<String, String> map = new HashMap<>();
+        Map<String, String> map = new HashMap<String, String>();
 
         Properties prop = loadI18nProp();
         if (keys!=null && keys.length>0) {
