@@ -11,8 +11,6 @@ import java.util.List;
  */
 public interface AdminBiz {
 
-    public static final String MAPPING = "/api";
-
 
     // ---------------------- callback ----------------------
 
@@ -42,5 +40,9 @@ public interface AdminBiz {
      * @return
      */
     public ReturnT<String> registryRemove(RegistryParam registryParam);
+
+
+    // ---------------------- biz (custome) ----------------------
+    // group、job ... manage
 
 }
