@@ -3,12 +3,13 @@ package com.xxl.job.core.thread;
 import com.xxl.job.core.biz.AdminBiz;
 import com.xxl.job.core.biz.model.HandleCallbackParam;
 import com.xxl.job.core.biz.model.ReturnT;
+import com.xxl.job.core.context.XxlJobContext;
+import com.xxl.job.core.context.XxlJobHelper;
 import com.xxl.job.core.enums.RegistryConfig;
 import com.xxl.job.core.executor.XxlJobExecutor;
 import com.xxl.job.core.log.XxlJobFileAppender;
-import com.xxl.job.core.log.XxlJobLogger;
 import com.xxl.job.core.util.FileUtil;
-import com.xxl.job.core.util.JacksonUtil;
+import com.xxl.job.core.util.JdkSerializeTool;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -76,7 +77,9 @@ public class TriggerCallbackThread {
                             }
                         }
                     } catch (Exception e) {
-                        logger.error(e.getMessage(), e);
+                        if (!toStop) {
+                            logger.error(e.getMessage(), e);
+                        }
                     }
                 }
 
@@ -88,13 +91,16 @@ public class TriggerCallbackThread {
                         doCallback(callbackParamList);
                     }
                 } catch (Exception e) {
-                    logger.error(e.getMessage(), e);
+                    if (!toStop) {
+                        logger.error(e.getMessage(), e);
+                    }
                 }
                 logger.info(">>>>>>>>>>> xxl-job, executor callback thread destory.");
 
             }
         });
         triggerCallbackThread.setDaemon(true);
+        triggerCallbackThread.setName("xxl-job, executor TriggerCallbackThread");
         triggerCallbackThread.start();
 
 
@@ -106,12 +112,17 @@ public class TriggerCallbackThread {
                     try {
                         retryFailCallbackFile();
                     } catch (Exception e) {
-                        logger.error(e.getMessage(), e);
+                        if (!toStop) {
+                            logger.error(e.getMessage(), e);
+                        }
+
                     }
                     try {
                         TimeUnit.SECONDS.sleep(RegistryConfig.BEAT_TIMEOUT);
                     } catch (InterruptedException e) {
-                        logger.warn(">>>>>>>>>>> xxl-job, executor retry callback thread interrupted, error msg:{}", e.getMessage());
+                        if (!toStop) {
+                            logger.error(e.getMessage(), e);
+                        }
                     }
                 }
                 logger.info(">>>>>>>>>>> xxl-job, executor retry callback thread destory.");
@@ -124,20 +135,25 @@ public class TriggerCallbackThread {
     public void toStop(){
         toStop = true;
         // stop callback, interrupt and wait
-        triggerCallbackThread.interrupt();
-        try {
-            triggerCallbackThread.join();
-        } catch (InterruptedException e) {
-            logger.error(e.getMessage(), e);
+        if (triggerCallbackThread != null) {    // support empty admin address
+            triggerCallbackThread.interrupt();
+            try {
+                triggerCallbackThread.join();
+            } catch (InterruptedException e) {
+                logger.error(e.getMessage(), e);
+            }
         }
 
         // stop retry, interrupt and wait
-        triggerRetryCallbackThread.interrupt();
-        try {
-            triggerRetryCallbackThread.join();
-        } catch (InterruptedException e) {
-            logger.error(e.getMessage(), e);
+        if (triggerRetryCallbackThread != null) {
+            triggerRetryCallbackThread.interrupt();
+            try {
+                triggerRetryCallbackThread.join();
+            } catch (InterruptedException e) {
+                logger.error(e.getMessage(), e);
+            }
         }
+
     }
 
     /**
@@ -172,54 +188,73 @@ public class TriggerCallbackThread {
     private void callbackLog(List<HandleCallbackParam> callbackParamList, String logContent){
         for (HandleCallbackParam callbackParam: callbackParamList) {
             String logFileName = XxlJobFileAppender.makeLogFileName(new Date(callbackParam.getLogDateTim()), callbackParam.getLogId());
-            XxlJobFileAppender.contextHolder.set(logFileName);
-            XxlJobLogger.log(logContent);
+            XxlJobContext.setXxlJobContext(new XxlJobContext(
+                    -1,
+                    null,
+                    logFileName,
+                    -1,
+                    -1));
+            XxlJobHelper.log(logContent);
         }
     }
 
 
     // ---------------------- fail-callback file ----------------------
 
-    private static String failCallbackFileName = XxlJobFileAppender.getLogPath().concat(File.separator).concat("xxl-job-callback").concat(".log");
+    private static String failCallbackFilePath = XxlJobFileAppender.getLogPath().concat(File.separator).concat("callbacklog").concat(File.separator);
+    private static String failCallbackFileName = failCallbackFilePath.concat("xxl-job-callback-{x}").concat(".log");
 
     private void appendFailCallbackFile(List<HandleCallbackParam> callbackParamList){
+        // valid
+        if (callbackParamList==null || callbackParamList.size()==0) {
+            return;
+        }
+
         // append file
-        String content = JacksonUtil.writeValueAsString(callbackParamList);
-        FileUtil.appendFileLine(failCallbackFileName, content);
+        byte[] callbackParamList_bytes = JdkSerializeTool.serialize(callbackParamList);
+
+        File callbackLogFile = new File(failCallbackFileName.replace("{x}", String.valueOf(System.currentTimeMillis())));
+        if (callbackLogFile.exists()) {
+            for (int i = 0; i < 100; i++) {
+                callbackLogFile = new File(failCallbackFileName.replace("{x}", String.valueOf(System.currentTimeMillis()).concat("-").concat(String.valueOf(i)) ));
+                if (!callbackLogFile.exists()) {
+                    break;
+                }
+            }
+        }
+        FileUtil.writeFileContent(callbackLogFile, callbackParamList_bytes);
     }
 
     private void retryFailCallbackFile(){
 
-        // load and clear file
-        List<String> fileLines = FileUtil.loadFileLines(failCallbackFileName);
-        FileUtil.deleteFile(failCallbackFileName);
-
-        // parse
-        List<HandleCallbackParam> failCallbackParamList = new ArrayList<>();
-        if (fileLines!=null && fileLines.size()>0) {
-            for (String line: fileLines) {
-                List<HandleCallbackParam> failCallbackParamListTmp = JacksonUtil.readValue(line, List.class, HandleCallbackParam.class);
-                if (failCallbackParamListTmp!=null && failCallbackParamListTmp.size()>0) {
-                    failCallbackParamList.addAll(failCallbackParamListTmp);
-                }
-            }
+        // valid
+        File callbackLogPath = new File(failCallbackFilePath);
+        if (!callbackLogPath.exists()) {
+            return;
+        }
+        if (callbackLogPath.isFile()) {
+            callbackLogPath.delete();
+        }
+        if (!(callbackLogPath.isDirectory() && callbackLogPath.list()!=null && callbackLogPath.list().length>0)) {
+            return;
         }
 
-        // retry callback, 100 lines per page
-        if (failCallbackParamList!=null && failCallbackParamList.size()>0) {
-            int pagesize = 100;
-            List<HandleCallbackParam> pageData = new ArrayList<>();
-            for (int i = 0; i < failCallbackParamList.size(); i++) {
-                pageData.add(failCallbackParamList.get(i));
-                if (i>0 && i%pagesize == 0) {
-                    doCallback(pageData);
-                    pageData.clear();
-                }
+        // load and clear file, retry
+        for (File callbaclLogFile: callbackLogPath.listFiles()) {
+            byte[] callbackParamList_bytes = FileUtil.readFileContent(callbaclLogFile);
+
+            // avoid empty file
+            if(callbackParamList_bytes == null || callbackParamList_bytes.length < 1){
+                callbaclLogFile.delete();
+                continue;
             }
-            if (pageData.size() > 0) {
-                doCallback(pageData);
-            }
+
+            List<HandleCallbackParam> callbackParamList = (List<HandleCallbackParam>) JdkSerializeTool.deserialize(callbackParamList_bytes, List.class);
+
+            callbaclLogFile.delete();
+            doCallback(callbackParamList);
         }
+
     }
 
 }
